Moved web samples entry point into their own module
- Hopefully should only have refactoring left.
