+++ samples/web/src/webMain/resources/index.html
@@ -53,6 +53,6 @@
 </head>
 <body>
 <div id="ComposeApp"></div>
-<script type="application/javascript" src="composeApp.js"></script>
+<script type="application/javascript" src="web.js"></script>
 </body>
 </html>
