+++ samples/web/webpack.config.d/watch.js
