+++ settings.gradle.kts
@@ -34,3 +34,4 @@ include(":core:testing")
 include(":samples:android")
 include(":samples:jvm")
 include(":samples:composeApp")
+include(":samples:web")
